Move state structs to a dedicated module
