+++ crates/starknet_api/src/lib.rs
@@ -16,8 +16,8 @@ pub use self::block::{
 pub use self::core::{ClassHash, ContractAddress, Nonce};
 pub use self::hash::{StarkFelt, StarkHash, GENESIS_HASH};
 pub use self::state::{
-    DeclaredContract, DeployedContract, IndexedDeclaredContract, IndexedDeployedContract,
-    StateDiff, StateNumber, StorageDiff, StorageEntry, StorageKey,
+    DeclaredContract, DeployedContract, StateDiff, StateNumber, StorageDiff, StorageEntry,
+    StorageKey,
 };
 pub use self::transaction::{
     CallData, ContractAddressSalt, ContractClass, DeclareTransaction, DeclareTransactionReceipt,
